+++ src/app/components/pages/TeacherRegistration.tsx
@@ -38,7 +38,7 @@ import { RegistrationNameInput } from "../elements/inputs/RegistrationNameInput"
 import { RegistrationDobInput } from "../elements/inputs/RegistrationDobInput";
 import { PasswordInputs } from "../elements/inputs/PasswordInput";
 import TeacherVerification from "../elements/inputs/TeacherVerification";
-import useRegistration from "../handlers/useRegistration";
+import useRegistration from "../hooks/useRegistration";
 import { RegistrationSubmit } from "../elements/inputs/RegistrationSubmit";
 import ReCAPTCHA from "react-google-recaptcha";
 import { Recaptcha } from "../elements/inputs/Recaptcha";
