+++ src/app/components/site/Homepage.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect } from "react";
 import { useAppSelector, selectors, isaacApi } from "../../state";
 import { Link } from "react-router-dom";
-import { Button, Col, Container, Row } from "reactstrap";
+import { Button, Col, Container } from "reactstrap";
 import { SITE_SUBJECT_TITLE } from "../../services";
 import { WhySignUpTabs } from "../elements/WhySignUpTabs";
 import { NewsCarousel } from "../elements/NewsCarousel";
@@ -9,199 +9,57 @@ import { FeaturedContentTabs } from "../elements/FeaturedContentTabs";
 import { EventsCarousel } from "../elements/EventsCarousel";
 import { FeaturedNewsItem } from "../elements/FeaturedNewsItem";
 import classNames from "classnames";
-import { TeacherPromoItem } from "../elements/TeacherPromoItem";
 import { PromoContent } from "../elements/PromoContent";
 import { ShowLoading } from "../handlers/ShowLoading";
+import { Dashboard } from "../elements/Dashboard";
 import { IsaacPodDTO } from "../../../IsaacApiTypes";
-
-interface ShowMeButtonsProps {
-  className?: string;
-}
 
 export const Homepage = () => {
   useEffect(() => {
     document.title = "Isaac " + SITE_SUBJECT_TITLE;
   }, []);
+
   const user = useAppSelector(selectors.user.orNull);
+
   const { data: news } = isaacApi.endpoints.getNewsPodList.useQuery({
     subject: "news",
     orderDescending: true,
   });
+
   const { data: promo } = isaacApi.endpoints.getNewsPodList.useQuery({
     subject: "promo",
     orderDescending: true,
   });
 
-  const featuredNewsItem = news && user?.loggedIn ? news[0] : undefined;
-
-  let carouselNewsItems: IsaacPodDTO[] = [];
-  if (news) {
-    if (user?.loggedIn) {
-      carouselNewsItems =
-        user?.role === "TEACHER" && promo ? news : news.slice(1);
-    } else carouselNewsItems = news;
-  }
-
   const promoItem = promo ? promo[0] : undefined;
 
-  const PromoOrFeaturedNews = ({
-    contentType,
-  }: {
-    contentType: "promo" | "news";
-  }) => {
-    const { dataTestId, className, content } = {
-      promo: {
-        dataTestId: "promo-tile",
-        className: "mt-4 d-block",
-        content: <TeacherPromoItem item={promoItem} />,
-      },
-      news: {
-        dataTestId: "featured-news-item",
-        className: "d-none d-lg-block",
-        content: <FeaturedNewsItem item={featuredNewsItem} />,
-      },
-    }[contentType];
+  const featuredNewsItem = (news && user?.loggedIn) ? news[0] : undefined;
 
-    return (
-      <Col md="12" lg="7" className={className} data-testid={dataTestId}>
-        {content}
-      </Col>
-    );
-  };
+  let carouselNewsItems: IsaacPodDTO[] = [];
 
-  const ShowMeButtons = ({ className }: ShowMeButtonsProps) => (
-    <Container
-      id="homepageButtons"
-      className={`${className} ${!user?.loggedIn ? "pt-0 px-lg-0" : ""}`}
-    >
-      <h3>Show me</h3>
-      <Row>
-        <Col xs={12} lg={user?.loggedIn ? 12 : 4} className="py-1">
-          <Button
-            size="lg"
-            tag={Link}
-            to={"/topics/gcse"}
-            color="secondary"
-            block
-          >
-            GCSE resources
-          </Button>
-        </Col>
-        <Col xs={12} lg={user?.loggedIn ? 12 : 4} className="py-1">
-          <Button
-            size="lg"
-            tag={Link}
-            to={"/topics/a_level"}
-            color="secondary"
-            block
-          >
-            A Level resources
-          </Button>
-        </Col>
-        <Col xs={12} lg={user?.loggedIn ? 12 : 4} className="py-1">
-          <Button size="lg" tag={Link} to={"/events"} color="secondary" block>
-            Events
-          </Button>
-        </Col>
-      </Row>
-    </Container>
-  );
+  if (news) {
+    (user?.loggedIn && user?.role === "TEACHER") || user?.loggedIn === false
+      ? (carouselNewsItems = news)
+      : (carouselNewsItems = news.slice(1));
+  }
 
   return (
     <>
       <div id="homepage">
         <section id="call-to-action" className="homepageHero">
           <Container className="pt-4 z1">
-            {user?.loggedIn ? (
-              <>
-                <Row className="pt-4">
-                  <Col md="12" lg="5" className={"pt-lg-4"}>
-                    <Container className={"mb-4"}>
-                      <h1 id="homepageName">Welcome {user.givenName}</h1>
-                    </Container>
-                    <ShowMeButtons className={"pt-xl-2"} />
-                    {/*<img id="homepageHeroImg" className="img-fluid" alt="Three Computer Science students studying with two laptops, one with code on the screen" src="/assets/ics_hero.svg" />*/}
-                  </Col>
-                  <PromoOrFeaturedNews
-                    contentType={
-                      user.role === "TEACHER" && promo ? "promo" : "news"
-                    }
-                  />
-                </Row>
-              </>
-            ) : (
-              <Row>
-                <Col lg="5" className="order-first pb-3">
-                  <Row>
-                    <Col>
-                      <h1>Computer science learning</h1>
-                      <p className="mt-4">
-                        Welcome to Isaac Computer Science, the free online
-                        platform for students and teachers.
-                      </p>
-                      <ul>
-                        <li>
-                          Use it in the <strong>classroom</strong>
-                        </li>
-                        <li>
-                          Use it for <strong>homework</strong>
-                        </li>
-                        <li>
-                          Use it for <strong>revision</strong>
-                        </li>
-                      </ul>
-
-                      <p className="mr-lg-n1">
-                        {"We also offer free "}
-                        <Link to="/events?types=teacher">
-                          teacher CPD events
-                        </Link>
-                        {" and "}
-                        <Link to="/events?types=student">
-                          student workshops
-                        </Link>
-                        .<br />
-                        {
-                          "Isaac Computer Science is proud to be part of the Department for Education's "
-                        }
-                        <Link to="/teachcomputing">
-                          National Centre for Computing Education
-                        </Link>
-                        .
-                      </p>
-                    </Col>
-                  </Row>
-                </Col>
-                <Col
-                  lg="7"
-                  className="order-last order-lg-1 px-lg-5 align-self-center text-center pattern-03"
-                >
-                  <iframe
-                    title="Isaac Computer Science introduction video"
-                    width="640"
-                    height="345"
-                    className="mw-100 pt-lg-4"
-                    src="https://www.youtube-nocookie.com/embed/ci6_Du_NHZA?enablejsapi=1&rel=0&fs=1&modestbranding=1&origin=home"
-                    frameBorder="0"
-                    allowFullScreen
-                  />
-                </Col>
-                <Col className="order-lg-last pb-5 pb-lg-3">
-                  <ShowMeButtons />
-                </Col>
-              </Row>
-            )}
+            <Dashboard
+              promoItem={promoItem}
+              featuredNewsItem={featuredNewsItem}
+            />
           </Container>
         </section>
 
         {!user?.loggedIn && (
-          <Container>
-            <hr />
-          </Container>
-        )}
-
-        {!user?.loggedIn && (
           <>
+            <Container>
+              <hr />
+            </Container>
             <section id="why-sign-up" className="row sign-up-tabs">
               <Container>
                 <Col className="pb-5 pt-4 pattern-04">
@@ -210,7 +68,6 @@ export const Homepage = () => {
                 </Col>
               </Container>
             </section>
-
             <section id="promo-content" className="row bg-primary pattern-05">
               <ShowLoading
                 until={promoItem}
@@ -237,8 +94,8 @@ export const Homepage = () => {
               className="eventList pt-5 pattern-03-reverse"
             >
               <h2 className="h-title mb-4">News</h2>
-              {user?.loggedIn && user?.role !== "TEACHER" && (
-                <div className={"d-block d-lg-none mb-4 mb-lg-0"}>
+              {user?.loggedIn && user.role !== "TEACHER" && (
+                <div className="d-block d-lg-none mb-4 mb-lg-0">
                   <FeaturedNewsItem item={featuredNewsItem} />
                 </div>
               )}
