+++ src/scss/common/media.scss
@@ -1,5 +1,9 @@
 // ISAAC
 
+iframe.no-border {
+  border: none;
+}
+
 figure {
   background-color: white;
   img {
